Move TestedCasesDetailsTable to components/index/cards/charts
